+++ tests/Fixtures/TestInvokableController.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Honeybadger\Tests\Fixtures;
+
+class TestInvokableController
+{
+    public function __invoke()
+    {
+        return response()->json([]);
+    }
+}